:white_check_mark: Complete hexagonal grid implementation and add tests
