+++ src/mesh_generator.jl
@@ -30,9 +30,10 @@ end
 
 @inline _mapToGlobalIdx(cells,cellidx,localvertexidx) = cells[cellidx].vertices[localvertexidx]
 
-function _get_vertexset_from_edges(cells,edgeset,CellType)
+function _get_vertexset_from_edges(cells,edgeset)
     vertices = Set{Int}()
     for edge in edgeset
+        CellType = typeof(cells[edge.cellidx])
         push!(vertices, _mapToGlobalIdx(cells, edge.cellidx, reference_edge_vertices(CellType)[edge.idx][1]))
         push!(vertices, _mapToGlobalIdx(cells, edge.cellidx, reference_edge_vertices(CellType)[edge.idx][2]))
     end
@@ -85,7 +86,7 @@ function rectangle_mesh(::Type{TriangleCell}, nel::NTuple{2,Int}, LL::Tensors.Ve
     edgesets["boundary"] = union(edgesets["bottom"],edgesets["right"],edgesets["top"],edgesets["left"])
     vertexsets = Dict{String,Set{Int}}()
     for set in edgesets
-        vertexsets[set.first] = _get_vertexset_from_edges(cells,set.second, TriangleCell)
+        vertexsets[set.first] = _get_vertexset_from_edges(cells,set.second)
     end
     return PolytopalMesh(cells, vertices; edgesets = edgesets, vertexsets = vertexsets)
 end
@@ -134,7 +135,139 @@ function rectangle_mesh(::Type{RectangleCell}, nel::NTuple{2,Int}, LL::Tensors.V
     edgesets["boundary"] = union(edgesets["bottom"],edgesets["right"],edgesets["top"],edgesets["left"])
     vertexsets = Dict{String,Set{Int}}()
     for set in edgesets
-        vertexsets[set.first] = _get_vertexset_from_edges(cells,set.second, RectangleCell)
+        vertexsets[set.first] = _get_vertexset_from_edges(cells,set.second)
+    end
+    return PolytopalMesh(cells, vertices; edgesets = edgesets, vertexsets = vertexsets)
+end
+
+#########################
+# Hexagonal Cells 2D
+#############################
+
+function _generate_2d_hex_centroids(T,LL, n_centroid_rows, n_centroid_cols, hex_width, hex_heigth)
+    centroids = Tensors.Vec{2,T}[]
+    x_coord = LL[1]; y_coord = LL[2]; sign = 1;
+    for j in 0:(n_centroid_rows-1)
+        for i in 0:(sign > 0 ? n_centroid_cols : n_centroid_cols-1)
+            centroid = Tensors.Vec{2,T}((x_coord,y_coord))
+            push!(centroids, centroid)
+            x_coord = x_coord + hex_width
+        end
+        y_coord = y_coord + hex_heigth*3/4
+        x_coord = LL[1] + max(0,sign*hex_width/2)
+        sign =  -sign;
+    end
+    return centroids
+end
+
+function _gen_hexagon(centroid,LL,UR,hex_width,hex_heigth)
+    hex_verts = [
+        Vertex(Tensors.Vec{2}((centroid[1]-hex_width/2,centroid[2]-hex_heigth/4))),
+        Vertex(Tensors.Vec{2}((centroid[1],centroid[2]-hex_heigth/2))),
+        Vertex(Tensors.Vec{2}((centroid[1]+hex_width/2,centroid[2]-hex_heigth/4))),
+        Vertex(Tensors.Vec{2}((centroid[1]+hex_width/2,centroid[2]+hex_heigth/4))),
+        Vertex(Tensors.Vec{2}((centroid[1],centroid[2]+hex_heigth/2))),
+        Vertex(Tensors.Vec{2}((centroid[1]-hex_width/2,centroid[2]+hex_heigth/4))),
+    ]
+    filter_verts = filter(x->(x.x[1]>=LL[1] && x.x[1]<=UR[1]
+                          && x.x[2]>=LL[2] && x.x[2] <= UR[2]),hex_verts)
+    if size(filter_verts,1) == 2 #corners case
+        x = centroid[1] + (centroid[1] == LL[1] ? hex_width/2 : -hex_width/2)
+        new_vertex = Vertex(Tensors.Vec{2}((x, centroid[2])))
+        if centroid == LL
+            filter_verts = (Vertex(centroid),new_vertex,filter_verts...)
+        elseif centroid[2] == LL[2] && centroid[1] == UR[1]
+            filter_verts = (new_vertex,Vertex(centroid),filter_verts...)
+        elseif centroid[2] == UR[2] && centroid[1] == LL[1]
+            filter_verts = (filter_verts...,new_vertex,Vertex(centroid))
+        elseif centroid == UR
+            filter_verts = (filter_verts...,Vertex(centroid),new_vertex)
+        else
+            throw("error on hexagon at corner $centroid")
+        end
+    elseif size(filter_verts,1) == 3 #top and bottom pentagon case
+        nv1 = Vertex(Tensors.Vec{2}((centroid[1]-hex_width/2, centroid[2])))
+        nv2 = Vertex(Tensors.Vec{2}((centroid[1]+hex_width/2, centroid[2])))
+        if centroid[2] == LL[2]
+            filter_verts = (nv1,nv2,filter_verts...)
+        elseif centroid[2] == UR[2]
+            filter_verts = (nv1,filter_verts...,nv2)
+        else
+            throw("error at bottom centroid $centroid")
+        end
+    end
+    return filter_verts
+end
+
+# Use map_func to avoid float precision differences between the same vertex
+function _push_cell_vertex!(vert,cell_verts, used_vertices,vertices,nextvert, map_func)
+        token = ht_keyindex2!(used_vertices, vert)
+        if token > 0 # reuse dofs
+            reuse_vert = used_vertices.vals[token]
+            push!(cell_verts, reuse_vert)
+        else # token <= 0, use new vertex
+            Base._setindex!(used_vertices, nextvert, vert, -token)
+            push!(cell_verts, nextvert)
+            push!(vertices, map_func(vert))
+            nextvert += 1
+        end
+        return nextvert
+end
+
+function rectangle_mesh(::Type{HexagonCell}, nel::NTuple{2,Int}, LL::Tensors.Vec{2,T}, UR::Tensors.Vec{2,T}) where {T}
+    LR = Tensors.Vec{2}((UR[1],LL[2]))
+    UL = Tensors.Vec{2}((LL[1],UR[2]))
+    nel_x = nel[1]; nel_y = isodd(nel[2]) ? nel[2] : nel[2]+1
+    nel_tot = 2*nel_x*nel_y +nel_y - nel_x
+
+    # Generate vertices
+    vertices = Vertex{2,T}[]
+    hex_width = 1.0
+    hex_heigth = 1.0
+    w_scale = (LR[1] - LL[1])/nel_x
+    h_scale = (UL[2] - LL[1])/nel_y
+    LLn = Tensors.Vec{2}((0.0,0.0))
+    URn = Tensors.Vec{2}((nel_x,nel_y))
+
+    map_func(x::Vertex) = Vertex(Tensors.Vec{2}((x.x[1]*w_scale+LL[1],x.x[2]*h_scale+LL[2])))
+
+    n_centroid_cols = nel_x; n_centroid_rows = nel_y+2
+    centroids = _generate_2d_hex_centroids(T,LL, n_centroid_rows, n_centroid_cols, hex_width, hex_heigth)
+    cells = Cell[]
+    used_vertices = Dict{Vertex,Int}()
+    #Add cells
+    nextvert = 1 # next free vertex to use
+    c_i = 1
+    sign = 1
+    for j in 0:(n_centroid_rows-1)
+        for i in 0:(sign > 0 ? n_centroid_cols : n_centroid_cols-1)
+            cell_verts = Int[]
+            for vert in _gen_hexagon(centroids[c_i],LLn,URn,hex_width,hex_heigth)
+                nextvert = _push_cell_vertex!(vert,cell_verts, used_vertices,vertices,nextvert, map_func)
+            end
+            n = size(cell_verts,1)
+            push!(cells,Cell{2,n,n,1}(Tuple(cell_verts)))
+            c_i +=1
+        end
+        sign =  -sign;
+    end
+
+    # Cell edges
+    edgesets = Dict{String,Set{EdgeIndex}}()
+    ncells = size(cells,1)
+    l_idxs = zip(repeat([6,4],div(nel_y,2)+1),cumsum(repeat([nel_x+1,nel_x],2)).+1)
+    ft_idx = sum(repeat([nel_x+1,nel_x],2)).+1
+    t_idxs = zip([3,repeat([4],nel_x-1)...,3],ft_idx:ft_idx+nel_x)
+    r_idxs = zip(repeat([3,2],div(nel_y,2)+1),cumsum(repeat([nel_x,nel_x+1],2)).+(nel_x+1))
+    edgesets["bottom"] = Set{EdgeIndex}([EdgeIndex(c_i,1) for c_i in 1:(nel_x+1)])
+    edgesets["right"]  = Set{EdgeIndex}([EdgeIndex(1+nel_x,2),[EdgeIndex(j,i) for (i,j) in r_idxs]...])
+    edgesets["top"]    = Set{EdgeIndex}([EdgeIndex(j,i) for (i,j) in t_idxs])
+    edgesets["left"]   = Set{EdgeIndex}([EdgeIndex(1,4),[EdgeIndex(j,i) for (i,j) in l_idxs]...])
+    edgesets["boundary"] = union(edgesets["bottom"],edgesets["right"],edgesets["top"],edgesets["left"])
+
+    vertexsets = Dict{String,Set{Int}}()
+    for set in edgesets
+        vertexsets[set.first] = _get_vertexset_from_edges(cells,set.second)
     end
     return PolytopalMesh(cells, vertices; edgesets = edgesets, vertexsets = vertexsets)
 end
